+++ rust/stackablectl/CHANGELOG.md
@@ -6,9 +6,10 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
-- Automatically detect Kubernetes environment (e.g. kind, k3s or IONOS) and choose a sensible [listener-operator preset] by default ([#XXX]).
-- Support configuring the [listener-operator preset] using `--listener-operator-preset` ([#XXX]).
+- Automatically detect Kubernetes environment (e.g. kind, k3s or IONOS) and choose a sensible [listener-operator preset] by default ([#414]).
+- Support configuring the [listener-operator preset] using `--listener-operator-preset` ([#414]).
 
+[#414]: https://github.com/stackabletech/stackable-cockpit/pull/414
 [listener-operator preset]: https://docs.stackable.tech/home/nightly/listener-operator/listenerclass/#presets
 
 ## [1.1.0] - 2025-07-16